Move `prepareOutputs` from `TransactionLayer` to `CoinSelection.Integrated`.
