+++ lib/core/src/Cardano/Wallet/Primitive/CoinSelection/Integrated.hs
@@ -1,11 +1,18 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedLabels #-}
 
 module Cardano.Wallet.Primitive.CoinSelection.Integrated
     ( performSelection
     , SelectionConstraints (..)
     , SelectionData (..)
+
+    , prepareOutputs
+    , ErrPrepareOutputs (..)
+    , ErrOutputTokenBundleSizeExceedsLimit (..)
+    , ErrOutputTokenQuantityExceedsLimit (..)
     ) where
 
 import Prelude
@@ -17,18 +24,30 @@ import Cardano.Wallet.Primitive.CoinSelection.Balanced
     , SelectionResult
     , SelectionSkeleton
     )
+import Cardano.Wallet.Primitive.Types.Address
+    ( Address )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin )
 import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle )
 import Cardano.Wallet.Primitive.Types.TokenMap
-    ( TokenMap )
+    ( AssetId, TokenMap )
+import Cardano.Wallet.Primitive.Types.TokenQuantity
+    ( TokenQuantity )
 import Cardano.Wallet.Primitive.Types.Tx
-    ( TokenBundleSizeAssessor, TxOut )
+    ( TokenBundleSizeAssessment (..)
+    , TokenBundleSizeAssessor (..)
+    , TxOut
+    , txOutMaxTokenQuantity
+    )
 import Cardano.Wallet.Primitive.Types.UTxOIndex
     ( UTxOIndex )
 import Control.Monad.Random.Class
     ( MonadRandom )
+import Data.Generics.Internal.VL.Lens
+    ( view )
+import Data.Generics.Labels
+    ()
 import Data.List.NonEmpty
     ( NonEmpty (..) )
 import GHC.Generics
@@ -37,6 +56,10 @@ import GHC.Stack
     ( HasCallStack )
 
 import qualified Cardano.Wallet.Primitive.CoinSelection.Balanced as Balanced
+import qualified Cardano.Wallet.Primitive.Types.TokenBundle as TokenBundle
+import qualified Cardano.Wallet.Primitive.Types.TokenMap as TokenMap
+import qualified Data.Foldable as F
+import qualified Data.Set as Set
 
 performSelection
     :: (HasCallStack, MonadRandom m)
@@ -94,5 +117,96 @@ data SelectionData = SelectionData
         :: !(Maybe Coin)
     , utxoAvailable
         :: !UTxOIndex
+    }
+    deriving (Eq, Generic, Show)
+
+prepareOutputs
+    :: TokenBundleSizeAssessor
+    -> (TokenMap -> Coin)
+    -> NonEmpty TxOut
+    -> Either ErrPrepareOutputs (NonEmpty TxOut)
+prepareOutputs tokenBundleSizeAssessor computeMinAdaQuantity outputsUnprepared
+    | (address, assetCount) : _ <- excessivelyLargeBundles =
+        Left $
+            -- We encountered one or more excessively large token bundles.
+            -- Just report the first such bundle:
+            ErrPrepareOutputsTokenBundleSizeExceedsLimit $
+            ErrOutputTokenBundleSizeExceedsLimit {address, assetCount}
+    | (address, asset, quantity) : _ <- excessiveTokenQuantities =
+        Left $
+            -- We encountered one or more excessive token quantities.
+            -- Just report the first such quantity:
+            ErrPrepareOutputsTokenQuantityExceedsLimit $
+            ErrOutputTokenQuantityExceedsLimit
+                { address
+                , asset
+                , quantity
+                , quantityMaxBound = txOutMaxTokenQuantity
+                }
+    | otherwise =
+        pure outputsToCover
+  where
+    -- The complete list of token bundles whose serialized lengths are greater
+    -- than the limit of what is allowed in a transaction output:
+    excessivelyLargeBundles :: [(Address, Int)]
+    excessivelyLargeBundles =
+        [ (address, assetCount)
+        | output <- F.toList outputsToCover
+        , let bundle = view #tokens output
+        , bundleIsExcessivelyLarge bundle
+        , let address = view #address output
+        , let assetCount = Set.size $ TokenBundle.getAssets bundle
+        ]
+
+      where
+        bundleIsExcessivelyLarge b = case assessSize b of
+            TokenBundleSizeWithinLimit -> False
+            OutputTokenBundleSizeExceedsLimit -> True
+          where
+            assessSize = view #assessTokenBundleSize tokenBundleSizeAssessor
+
+    -- The complete list of token quantities that exceed the maximum quantity
+    -- allowed in a transaction output:
+    excessiveTokenQuantities :: [(Address, AssetId, TokenQuantity)]
+    excessiveTokenQuantities =
+        [ (address, asset, quantity)
+        | output <- F.toList outputsToCover
+        , let address = view #address output
+        , (asset, quantity) <-
+            TokenMap.toFlatList $ view #tokens $ view #tokens output
+        , quantity > txOutMaxTokenQuantity
+        ]
+
+    outputsToCover =
+        Balanced.prepareOutputsWith computeMinAdaQuantity outputsUnprepared
+
+-- | Indicates a problem when preparing outputs for a coin selection.
+data ErrPrepareOutputs
+    = ErrPrepareOutputsTokenBundleSizeExceedsLimit
+        ErrOutputTokenBundleSizeExceedsLimit
+    | ErrPrepareOutputsTokenQuantityExceedsLimit
+        ErrOutputTokenQuantityExceedsLimit
+    deriving (Eq, Generic, Show)
+
+data ErrOutputTokenBundleSizeExceedsLimit = ErrOutputTokenBundleSizeExceedsLimit
+    { address :: !Address
+      -- ^ The address to which this token bundle was to be sent.
+    , assetCount :: !Int
+      -- ^ The number of assets within the token bundle.
+    }
+    deriving (Eq, Generic, Show)
+
+-- | Indicates that a token quantity exceeds the maximum quantity that can
+--   appear in a transaction output's token bundle.
+--
+data ErrOutputTokenQuantityExceedsLimit = ErrOutputTokenQuantityExceedsLimit
+    { address :: !Address
+      -- ^ The address to which this token quantity was to be sent.
+    , asset :: !AssetId
+      -- ^ The asset identifier to which this token quantity corresponds.
+    , quantity :: !TokenQuantity
+      -- ^ The token quantity that exceeded the bound.
+    , quantityMaxBound :: !TokenQuantity
+      -- ^ The maximum allowable token quantity.
     }
     deriving (Eq, Generic, Show)
